third_party/WebKit/LayoutTests/inspector/console/paintworklet-console-selector-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected identifier
Tests console execution context selector.

Console context selector:
* top 
  ____PaintWorkletGlobalScope 
